+++ src/wai/annotations/core/_Writer.py
@@ -1,13 +1,17 @@
 from abc import abstractmethod
 from typing import Generic, Iterable
 
+from .logging import LoggingEnabled, StreamLogger
+from ._ImageInfo import ImageInfo
 from ._typing import ExternalFormat
 
 
-class Writer(Generic[ExternalFormat]):
+class Writer(LoggingEnabled, Generic[ExternalFormat]):
     """
     Base class for classes which can write a specific external format to disk.
     """
+    logger_name = "wai.annotations.writer"
+
     def __init__(self, output: str):
         super().__init__()
 
@@ -20,7 +24,11 @@ class Writer(Generic[ExternalFormat]):
 
         :param instances:   The instances to write to disk.
         """
-        self.write(instances, self.output)
+        # Create a stream processor to log when we are writing a file
+        stream_log = StreamLogger(self.logger().info,
+                                  lambda instance: f"Saving annotations for {self.extract_image_info_from_external_format(instance).filename}").process
+
+        self.write(stream_log(instances), self.output)
 
     @abstractmethod
     def write(self, instances: Iterable[ExternalFormat], path: str):
@@ -31,5 +39,15 @@ class Writer(Generic[ExternalFormat]):
         :param path:        The path to write the instances to. Can be a directory
                             or a specific filename depending on the disk-format
                             of the external format.
+        """
+        pass
+
+    @abstractmethod
+    def extract_image_info_from_external_format(self, instance: ExternalFormat) -> ImageInfo:
+        """
+        Extracts an image-info object from the external format of this writer.
+
+        :param instance:    The instance being written.
+        :return:            The image info for the instance.
         """
         pass