+++ src/wai/annotations/tf/__init__.py
@@ -3,6 +3,7 @@ Package for utility functions for working with Tensorflow.
 """
 from ._extract_feature import extract_feature
 from ._format import TensorflowExampleExternalFormat
+from ._image_info_from_example import image_info_from_example
 from ._LabelMapAccumulator import LabelMapAccumulator
 from ._make_feature import make_feature
 from ._mask_from_polygon import mask_from_polygon